Add main program "demo"
This is a simplified example to highlight a problem mixing python and
ruby, which is specific to g++/Linux. The issue has so far only been
uncovered after calling APIs that return a std::vector of things.

A new structure `Foo` has been introduced to return a vector of `Bar`
instances. This does not use the Swig director feature and will
demonstrate the issue if compiled with g++. Mysteriously if this is
compiled with clang, the issue disappears.